% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_data.R
\name{env}
\alias{env}
\title{Environmental Data}
\format{
A data frame with [number of rows] rows and [number of columns] columns.
  - Each row represents a sample.
  - Each column represents an environmental variable.
}
\source{
Example data created for demonstration purposes.
}
\description{
This dataset contains environmental variables that can be used to explain variations in microbiome data.
}
\examples{
data(env)
}
