% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_data.R
\name{otu}
\alias{otu}
\title{OTU/ASV Data}
\format{
A data frame with [number of rows] rows and [number of columns] columns.
  - Each row represents a sample.
  - Each column represents an OTU or ASV.
}
\source{
Example data created for demonstration purposes.
}
\description{
This dataset contains OTU (Operational Taxonomic Unit) or ASV (Amplicon Sequence Variant) data.
It is typically used for microbiome analysis.
}
\examples{
data(otu)
}
